Use relative time interval instead of casting to date in Denver
In order to look back X days from the current time (instead of the current date) we need to compare based on time, not date.

This change multiples the number of days passed into the function by 24 and uses the number of hours to determine the time window to query.

This should give us accurate DURU throughout the day, instead of resetting the start of the window to midnight denver time, every day.